file_interface --> file_io + flie_interface
