+++ src/cryptext/interfaces/__init__.py
@@ -5,18 +5,8 @@ from .password_interface import (
     encrypt_password,
     decrypt_password,
 )
-from .file_interface import PathFilterFunction, InvalidFileError
+from .file_io import PathFilterFunction, InvalidFileError
 from .file_interface import (
-    is_dir,
-    is_file,
-    exists,
-    join,
-    list_dir,
-    get_file_base_name,
-    create_file,
-    remove_file,
-    read_binary_file,
-    write_to_binary_file,
     get_password_path,
     get_plugin_path,
     list_plugins,
@@ -26,16 +16,6 @@ from .file_interface import (
 __all__ = [
     'PathFilterFunction',
     'InvalidFileError',
-    'is_dir',
-    'is_file',
-    'exists',
-    'join',
-    'list_dir',
-    'get_file_base_name',
-    'create_file',
-    'remove_file',
-    'read_binary_file',
-    'write_to_binary_file',
     'list_plugins',
     'list_passwords',
     'get_password_path',
